pyproject.toml
[project]
name = "flipkart-recommender"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datasets>=4.5.0",
    "flask>=3.1.2",
    "langchain>=1.2.10",
    "langchain-astradb>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.0",
    "pandas>=3.0.0",
    "pypdf>=6.7.0",
    "python-dotenv>=1.2.1",
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
